:bug: Fix pre-release incrementing with numeric identifiers
If a version like `1.2.3-2022` was incremented with the specified pre-release 2022, `1.2.3-2023` would be returned. In this case, the user is clearly expecting a `0` to be added to the string, instead of incrementing it to `2023`, which wouldn't make sense.
This commit makes sure that the specified identifier is not incremented in these cases.